@@ -1,6 +1,6 @@
 {
     "name": "classifai",
-    "version": "0.0.0",
+    "version": "2.0.0",
     "scripts": {
         "format:prettier": "./node_modules/.bin/prettier --write src/**/*.{js,ts,json,css,scss,md}",
         "start:dev": "nodemon -w ./webpack.config.js --exec \"ng s --hmr --extra-webpack-config webpack.config.js\"",
@@ -15,15 +15,16 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "~11.0.2",
-        "@angular/cdk": "^12.2.2",
-        "@angular/common": "~11.0.2",
-        "@angular/compiler": "~11.0.2",
-        "@angular/core": "~11.0.2",
-        "@angular/forms": "~11.0.2",
-        "@angular/platform-browser": "~11.0.2",
-        "@angular/platform-browser-dynamic": "~11.0.2",
-        "@angular/router": "~11.0.2",
+        "@angular/animations": "~12.2.11",
+        "@angular/cdk": "^12.2.11",
+        "@angular/common": "~12.2.11",
+        "@angular/compiler": "~12.2.11",
+        "@angular/core": "~12.2.11",
+        "@angular/forms": "~12.2.11",
+        "@angular/platform-browser": "~12.2.11",
+        "@angular/platform-browser-dynamic": "~12.2.11",
+        "@angular/router": "~12.2.11",
+        "@generic-ui/ngx-grid": "^0.16.4",
         "@ngx-translate/core": "^13.0.0",
         "@ngx-translate/http-loader": "^6.0.0",
         "@swimlane/ngx-charts": "^19.0.1",
@@ -31,17 +32,19 @@
         "file-saver": "^2.0.5",
         "jszip": "^3.5.0",
         "lodash-es": "^4.17.15",
-        "ngx-build-plus": "^10.1.1",
+        "ngx-build-plus": "^10.1.0",
+        "papaparse": "^5.3.1",
         "rxjs": "~6.6.3",
         "streamsaver": "^2.0.4",
-        "tslib": "^2.0.0",
-        "zone.js": "~0.10.2"
+        "tslib": "^2.3.1",
+        "zone.js": "~0.11.4"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~0.1100.2",
-        "@angular/cli": "^11.0.2",
-        "@angular/compiler-cli": "~11.0.2",
-        "@angular/language-service": "~11.0.2",
+        "@angular-devkit/build-angular": "~12.2.11",
+        "@angular/cli": "^12.2.11",
+        "@angular/compiler-cli": "~12.2.11",
+        "@angular/language-service": "~12.2.11",
+        "@types/d3": "^7.1.0",
         "@types/file-saver": "^2.0.2",
         "@types/jasmine": "~3.6.0",
         "@types/jasminewd2": "~2.0.3",
@@ -49,9 +52,9 @@
         "@types/node": "^12.11.1",
         "codelyzer": "^6.0.0",
         "husky": "^4.2.5",
-        "jasmine-core": "~3.6.0",
+        "jasmine-core": "~3.8.0",
         "jasmine-spec-reporter": "~5.0.0",
-        "karma": "~5.0.0",
+        "karma": "~6.3.0",
         "karma-chrome-launcher": "~3.1.0",
         "karma-coverage-istanbul-reporter": "~3.0.2",
         "karma-jasmine": "~4.0.0",
@@ -61,7 +64,7 @@
         "protractor": "~7.0.0",
         "ts-node": "~8.3.0",
         "tslint": "~6.1.0",
-        "typescript": "~4.0.5",
+        "typescript": "~4.2.3",
         "webpack-bundle-analyzer": "^4.1.0"
     },
     "husky": {
