@@ -4,13 +4,13 @@
  * found in the LICENSE file at https://github.com/CertifaiAI/Classifai_FrontEnd/blob/main/LICENSE
  */
 
-import { distinctUntilChanged } from 'rxjs/operators';
+import { distinctUntilChanged, first } from 'rxjs/operators';
 import { environment } from 'environments/environment.prod';
-import { LabelList, Labels, Project, Folder } from '../../shared/types/dataset-layout/data-set-layout.model';
+import { LabelList, Labels, Project, Folder, Tabular } from '../../shared/types/dataset-layout/data-set-layout.model';
 import { HttpClient } from '@angular/common/http';
 import { ImageLabellingModeService } from 'components/image-labelling/image-labelling-mode.service';
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
+import { Observable, Subject } from 'rxjs';
 import { Router } from '@angular/router';
 import {
     Message,
@@ -23,17 +23,20 @@ import {
 } from 'shared/types/message/message.model';
 import { UnsupportedImageService } from 'shared/services/unsupported-image.service';
 import { ImageLabellingMode, BboxMetadata, PolyMetadata } from 'shared/types/image-labelling/image-labelling.model';
+import { LabellingModeService } from '../../shared/services/labelling-mode-service';
 
 @Injectable({ providedIn: 'any' })
 export class DataSetLayoutService {
     private hostPort: string = environment.baseURL;
     private imageLabellingMode: ImageLabellingMode = null;
+    private labellingMode: string | null = null;
 
     constructor(
         private http: HttpClient,
         private mode: ImageLabellingModeService,
         private router: Router,
         private _unsupportedImageService: UnsupportedImageService,
+        private _labellingModeService: LabellingModeService,
     ) {
         // if has mode value, acquire the mode value
         // else return to lading page
@@ -44,10 +47,14 @@ export class DataSetLayoutService {
                 this.router.navigate(['/']);
             }
         });
+
+        this._labellingModeService.labelMode$.pipe(distinctUntilChanged()).subscribe((labellingMode) => {
+            this.labellingMode = labellingMode;
+        });
     }
 
     getProjectList = (): Observable<MessageContent<Project[]>> => {
-        return this.http.get<MessageContent<Project[]>>(`${this.hostPort}${this.imageLabellingMode}/projects/meta`);
+        return this.http.get<MessageContent<Project[]>>(`${this.hostPort}${this.labellingMode}/projects/meta`);
         // .pipe(catchError(this.handleError));
     };
 
@@ -55,14 +62,24 @@ export class DataSetLayoutService {
         projectName: string,
         labelPath: string,
         projectFolderPath: string,
+        tabularFilePath: string,
     ): Observable<ProjectMessage> => {
-        const annotationType = this.imageLabellingMode === 'bndbox' ? 'boundingbox' : 'segmentation';
+        console.log(this.labellingMode);
+        let annotationType = '';
+        if (this.labellingMode === 'tabular') {
+            annotationType = 'tabular';
+        } else if (this.labellingMode === 'bndbox') {
+            annotationType = 'boundingbox';
+        } else if (this.labellingMode === 'seg') {
+            annotationType = 'segmentation';
+        }
         return this.http.put<ProjectMessage>(`${this.hostPort}v2/projects`, {
             project_name: projectName,
             annotation_type: annotationType,
             status: 'raw',
             project_path: projectFolderPath,
             label_file_path: labelPath,
+            tabular_file_path: tabularFilePath,
         });
     };
 
@@ -84,25 +101,23 @@ export class DataSetLayoutService {
     };
 
     updateProjectLoadStatus = (projectName: string): Observable<Message> => {
-        return this.http.get<Message>(`${this.hostPort}${this.imageLabellingMode}/projects/${projectName}`);
+        return this.http.get<Message>(`${this.hostPort}${this.labellingMode}/projects/${projectName}`);
     };
 
     checkProjectStatus = (projectName: string): Observable<MessageContent<Project[]>> => {
         return this.http.get<MessageContent<Project[]>>(
-            `${this.hostPort}${this.imageLabellingMode}/projects/${projectName}/meta`,
+            `${this.hostPort}${this.labellingMode}/projects/${projectName}/meta`,
         );
     };
 
     manualCloseProject = (projectName: string, status = 'closed'): Observable<Message> => {
-        return this.http.put<Message>(`${this.hostPort}${this.imageLabellingMode}/projects/${projectName}`, {
+        return this.http.put<Message>(`${this.hostPort}${this.labellingMode}/projects/${projectName}`, {
             status,
         });
     };
 
     checkExistProjectStatus = (projectName: string): Observable<LabelList> => {
-        return this.http.get<LabelList>(
-            `${this.hostPort}${this.imageLabellingMode}/projects/${projectName}/loadingstatus`,
-        );
+        return this.http.get<LabelList>(`${this.hostPort}${this.labellingMode}/projects/${projectName}/loadingstatus`);
     };
 
     getThumbnailList = (projectName: string, uuid: string): Observable<BboxMetadata & PolyMetadata> => {
@@ -112,9 +127,7 @@ export class DataSetLayoutService {
     };
 
     localUploadStatus = (projectName: string): Observable<MessageUploadStatus> => {
-        return this.http.get<MessageUploadStatus>(
-            `${this.hostPort}v2/${this.imageLabellingMode}/projects/${projectName}`,
-        );
+        return this.http.get<MessageUploadStatus>(`${this.hostPort}v2/${this.labellingMode}/projects/${projectName}`);
     };
 
     updateLabelList = (projectName: string, label_list: string[]): Observable<Message> => {
@@ -139,7 +152,7 @@ export class DataSetLayoutService {
     };
 
     importStatus = (): Observable<ImportResponse> => {
-        return this.http.get<ImportResponse>(`${this.hostPort}v2/${this.imageLabellingMode}/projects/importstatus`);
+        return this.http.get<ImportResponse>(`${this.hostPort}v2/${this.labellingMode}/projects/importstatus`);
     };
 
     importLabelFile() {
@@ -156,6 +169,14 @@ export class DataSetLayoutService {
 
     importProjectFolderStatus() {
         return this.http.get<Folder>(`${this.hostPort}v2/folders`);
+    }
+
+    importTabularFile() {
+        return this.http.put<Message>(`${this.hostPort}v2/tabularfile`, {});
+    }
+
+    importTabularFileStatus() {
+        return this.http.get<Tabular>(`${this.hostPort}v2/tabularfile`);
     }
 
     downloadUnsupportedImageList(projectName: string, unsupportedImageList: string[]) {
